fix: change resolution to resolve
* updated class to adopt the naming of ConflictResolve.default
* removed the ConflictResolution class, which we copied to ConflictResolver class.
* swift we named the protocol with a suffix, and objc kept same name for both.
* updated tests with chnages
