--- Swift/ConflictResolution.swift
@@ -1,49 +1,0 @@
-//
-//  ConflictResolution.swift
-//  CouchbaseLite
-//
-//  Copyright (c) 2019 Couchbase, Inc All rights reserved.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-
-import Foundation
-
-/// Conflict Resolution class
-public class ConflictResolution {
-    
-    /// The default conflict resolver used by the replicator.
-    public static var `default`: ConflictResolver {
-        return DefaultResolver()
-    }
-    
-}
-
-/* internal */ class DefaultResolver: ConflictResolver {
-    
-    let resolver: CBLConflictResolver
-    
-    init() {
-        resolver = CBLConflictResolution.default()
-    }
-    
-    func resolve(conflict: Conflict) -> Document? {
-        guard let doc = resolver.resolve(conflict.impl) else {
-            return nil
-        }
-        
-        return Document(doc)
-    }
-    
-}
-
